pyproject.toml
[project]
name = "zeroband"
version = "0.1.0"
description = "ZeroBand is a production ready codebase for decentralized training of LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "numpy",
    "setuptools",
    "transformers>=4.51.0",
    "datasets>=3.0.0",
    "torchdata>=0.8.0",
    "ninja",
    "zstandard",
    "pyarrow",
    "wandb",
    "vllm~=0.8.5",
    "jaxtyping",
    "beartype",
    "toploc>=0.1.6",
    "llmcompressor",
    "pylatexenc>=2.10",
    "liger_kernel",
    "google-cloud-storage",
    "shardcast >= 0.3.1",
    "asyncio>=3.4.3",
    "aiohttp>=3.10.5",
    "pyext @ git+https://github.com/justusmattern27/PyExt.git",
    "prime-iroh>=0.3.1",
    "setuptools",
    "fastapi",
    "uvicorn",
    "requests",
    "reasoning-gym @ git+https://github.com/open-thought/reasoning-gym.git",
    "tomli>=2.2.1",
    "pydantic-settings>=2.9.1",
    "cydifflib>=1.2.0",
] 


[project.optional-dependencies]
fa = ["flash-attn>=2.7.4"]
kernelbench = ["modal>=1.0.3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true # allow direct references to git repos in dependencies

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["F", "I"]
ignore = ["F722", "F821"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.0",
    "pre-commit>=3.0.0",
    "pytest>=7.0.0",
    "faker",
    "ipykernel>=6.29.5",
]
no-build-isolation-package = ["flash-attn"]

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as gpu (deselect with '-m \"not gpu\"')",
]
